PowerControl: Decompose `MenuStack.cs` into `Options/`
